+++ doc/components/component/data/plugin/upload.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="doc">
-    <h2>Upload 上传</h2>
+    <h2>Uploader 上传</h2>
     <p>HeyUI的上传，将提供一组样式，并提供七牛上传的demo。</p>
     <p>不同的系统可以根据自己的上传实现来开发。</p>
     <p>所有的示例将使用<a href="https://github.com/moxiecode/plupload" target="_blank">plupload</a>配套使用，因为组件已经将拖拽以及上传功能实现，这里也不做扩展了。</p>
@@ -16,7 +16,6 @@
 
     <h3>单图片上传-url</h3>    
     <example demo="dataplugins/uploader3"></example>
-
   
     <h3>多图片上传-url数组</h3>
     <example demo="dataplugins/uploader4"></example>
@@ -27,9 +26,11 @@
     <h3>多文件上传</h3>
     <example demo="dataplugins/uploader6"></example>
 
-  
     <h3>拖拽上传</h3>
     <example demo="dataplugins/uploader7"></example>
+
+    <h3>只读展示</h3>
+    <example demo="dataplugins/uploader8"></example>
 
 
     <h3>Uploader 参数</h3>
@@ -76,10 +77,37 @@
         <td>-</td>
         <td>-</td>
       </tr>
+      <tr>
+        <td>readonly</td>
+        <td>readonly, 1.5.0+</td>
+        <td>Boolean</td>
+        <td>-</td>
+        <td>false</td>
+      </tr>
     </table>
 
-    <h3 class="democodes">七牛的通用示例</h3>
-    <p>本示例写明了通用的处理方案，如果使用其他上传方式，请参照该文件修改。<a href="https://github.com/heyui/heyui/blob/master/doc/components/demos/common/qiniu.vue" target="_blank">文件链接</a></p>
+
+    <h3>Uploader 事件</h3>
+    <table class="table">
+      <tr>
+        <th>事件</th>
+        <th>说明</th>
+        <th>返回数据</th>
+      </tr>
+      <tr>
+        <td>fileClick</td>
+        <td>文件链接点击的时候</td>
+        <td>file, index</td>
+      </tr>
+    </table>
+
+    <h3 class="democodes">上传的通用示例</h3>
+    <p>本示例写明了通用的上传处理方案，如果使用自己服务器上传的方式，请参照该文件修改。<a href="https://github.com/heyui/heyui/blob/master/doc/components/demos/common/qiniu.vue" target="_blank">文件链接</a></p>
+    <codes src="/codes/uploader.txt" type="javascript"></codes>
+
+
+    <h3 class="democodes">七牛上传的通用示例</h3>
+    <p>本示例写明了七牛上传的处理方案。<a href="https://github.com/heyui/heyui/blob/master/doc/components/demos/common/qiniu.vue" target="_blank">文件链接</a></p>
     <codes src="/common/qiniu.vue" type="html"></codes>
   </div>
 </template>