+++ doc/components_en/common/frame.vue
@@ -39,7 +39,7 @@
     </div>
     <div class="right-frame h-dropdown-common-container">
       <router-view></router-view>
-      <footer class="com-frame-footer">HeyUI ©2019 Created by <a href="http://www.ch-un.com" target="_blank">vvpvvp</a></footer>
+      <footer class="com-frame-footer">Copyright © 2019 <a href="http://www.ch-un.com" target="_blank">Lan</a></footer>
       <BackTop :target="getTarget"></BackTop>
       <div class="right-frame-menu-container">
         <ul class="right-frame-menu" v-if="menus.length>0">