+++ doc/components_en/component/data/plugin/taginput.vue
@@ -56,5 +56,24 @@
         <td>Array</td>
       </tr>
     </table>
+
+    <h3>TagInput Event</h3>
+    <table class="table">
+      <tr>
+        <th>Event</th>
+        <th>Description</th>
+        <th>Return Value</th>
+      </tr>
+      <tr>
+        <td>input</td>
+        <td>When data changed</td>
+        <td>v-model value</td>
+      </tr>
+      <tr>
+        <td>change</td>
+        <td>When data changed</td>
+        <td>total object value</td>
+      </tr>
+    </table>
   </div>
 </template>