+++ doc/components_en/component/data/view/table.vue
@@ -11,6 +11,11 @@
     <example demo="view/table2"></example>
 
     <h3>Data render</h3>
+    <ul class="text-ul">
+      <li><code>unit</code>: setting data unit</li>
+      <li><code>dict</code>: use dict render data</li>
+      <li><code>render</code>: define render function</li>
+    </ul>
     <example demo="view/table9"></example>
 
     <h3>Sort and loading</h3>
@@ -102,10 +107,24 @@
         <td>-</td>
         <td>false</td>
       </tr>
+      <tr>
+        <td>selectWhenClickTr</td>
+        <td>When clicking on tr, trigger the checkbox switch</td>
+        <td>Boolean</td>
+        <td></td>
+        <td>false</td>
+      </tr>
+      <tr>
+        <td>sortProp</td>
+        <td>sort prop，when prop is null，use sortProp, <code>1.15.1+</code></td>
+        <td>String</td>
+        <td></td>
+        <td></td>
+      </tr>
     </table>
 
 
-    <h3>Column Property</h3>
+    <h3>TableItem / Column Property</h3>
     <table class="table">
       <tr>
         <th>Property</th>
@@ -142,7 +161,6 @@
         <td>-</td>
         <td></td>
       </tr>
-
       <tr>
         <td>align</td>
         <td>Header alignment in table</td>
@@ -179,11 +197,25 @@
         <td>false</td>
       </tr>
       <tr>
-        <td>selectWhenClickTr</td>
-        <td>When clicking on tr, trigger the checkbox switch</td>
-        <td>Boolean</td>
+        <td>dict</td>
+        <td>use dict render data</td>
+        <td>String</td>
         <td></td>
-        <td>false</td>
+        <td></td>
+      </tr>
+      <tr>
+        <td>unit</td>
+        <td>setting data unit</td>
+        <td>String</td>
+        <td></td>
+        <td></td>
+      </tr>
+      <tr>
+        <td>render</td>
+        <td>define render function</td>
+        <td>Function</td>
+        <td></td>
+        <td></td>
       </tr>
     </table>
 
@@ -192,30 +224,37 @@
       <tr>
         <th>Event</th>
         <th>Description</th>
+        <th>Return Value</th>
       </tr>
       <tr>
         <td>sort</td>
         <td>Events fired when sorted</td>
+        <td>{type: [asc|desc], prop: String}</td>
       </tr>
       <tr>
         <td>select</td>
         <td>Events that fire when checkbox is changed</td>
+        <td>checks</td>
       </tr>
       <tr>
         <td>selectAll</td>
         <td>When checkbox is selected</td>
+        <td>checks</td>
       </tr>
       <tr>
         <td>trclick</td>
         <td>When tr is clicked</td>
+        <td>(data, event, index)</td>
       </tr>
       <tr>
         <td>trdblclick</td>
         <td>When tr is double clicked</td>
+        <td>(data, event, index)</td>
       </tr>
       <tr>
         <td>rowSelect</td>
         <td>When tr is selected</td>
+        <td>data</td>
       </tr>
     </table>
 
@@ -224,54 +263,47 @@
       <tr>
         <th>Method</th>
         <th>Description</th>
+        <th>Parameter</th>
+        <th>Return Value</th>
       </tr>
       <tr>
         <td>clearSort</td>
         <td>Clear sort</td>
+        <td>Empty</td>
+        <td>Empty</td>
       </tr>
       <tr>
         <td>clearSelection</td>
         <td>Clear selected</td>
+        <td>Empty</td>
+        <td>Empty</td>
       </tr>
       <tr>
         <td>setSelection</td>
         <td>Set the selected value</td>
+        <td>data</td>
+        <td>Empty</td>
       </tr>
       <tr>
         <td>getSelection</td>
         <td>get selected values</td>
+        <td>Empty</td>
+        <td>checks: Array</td>
       </tr>
       <tr>
-        <td>inverseSelection</td>
+        <td>invereSelection</td>
         <td>inverse selection</td>
+        <td>Empty</td>
+        <td>Empty</td>
+      </tr>
+      <tr>
+        <td>triggerSort</td>
+        <td>Set sort, trigger sort event when triggerType is true, and sort automatically when triggerType is auto, <code>1.15.1+</code></td>
+        <td>sortStatus: {prop: String, type: String['asc'|'desc'] }, triggerType: true | 'auto' | false</td>
+        <td>Empty</td>
       </tr>
     </table>
 
-    <h3>TreePicker Event</h3>
-
-    <h3>Tree Event</h3>
-    <table class="table">
-      <tr>
-        <th>Event</th>
-        <th>Description</th>
-      </tr>
-      <tr>
-        <td>input</td>
-        <td>The v-model value changes</td>
-      </tr>
-      <tr>
-        <td>choose</td>
-        <td>Checkbox selection change</td>
-      </tr>
-      <tr>
-        <td>select</td>
-        <td>Tree's select change</td>
-      </tr>
-      <tr>
-        <td>loadDataSuccess</td>
-        <td>When the asynchronous data load is successful.	</td>
-      </tr>
-    </table>
   </div>
 </template>