+++ src/components/imagepreview/imagepreview.vue
@@ -1,0 +1,42 @@
+<template>
+  <Modal :value="value" @input="updateValue" className="h-image-preview-modal">
+    <Preview :isShow="value" :datas="datas" :index="index"></Preview>
+  </Modal>
+</template>
+<script>
+
+import utils from '../../utils/utils';
+import Modal from '../modal/modal';
+import Preview from './preview.vue';
+
+export default {
+  name: 'hImagePreviewModal',
+  props: {
+    value: {
+      type: Boolean,
+      default: false
+    },
+    datas: {
+      type: [Array],
+      default: () => ([])
+    },
+    index: {
+      type: Number,
+      default: 0
+    }
+  },
+  data() {
+    return {
+    }
+  },
+  methods: {
+    updateValue(value) {
+      this.$emit('input', value);
+    }
+  },
+  components: {
+    Modal,
+    Preview
+  }
+};
+</script>